+++ packages/react-core/src/deprecated/components/Chip/__tests__/ChipGroup.test.tsx
@@ -7,6 +7,8 @@ import { ChipGroup } from '../index';
 import { Chip } from '..';
 import styles from '@patternfly/react-styles/css/components/Label/label-group';
 
+jest.mock('../../../../helpers/GenerateId/GenerateId');
+
 test('chip group default', () => {
   render(
     <ChipGroup>
@@ -71,7 +73,7 @@ test('chip group has aria-labelledby attribute', () => {
       <Chip>1.1</Chip>
     </ChipGroup>
   );
-  expect(screen.getByRole('list')).toHaveAttribute('aria-labelledby', expect.stringContaining(`pf-random-id`));
+  expect(screen.getByRole('list')).toHaveAttribute('aria-labelledby', expect.stringContaining(`generated-id`));
 });
 
 test('chip group has aria-labelledby attribute set to ID value', () => {
@@ -266,7 +268,7 @@ test('passing defaultIsOpen of true causes the chip group to be expanded by defa
 
 test('chip group will not render if no children passed', () => {
   render(<ChipGroup />);
-  expect(screen.queryByRole('group')).toBeNull();
+  expect(screen.queryByRole('list')).toBeNull();
 });
 
 test('chip group renders to match snapshot', () => {