+++ packages/react-templates/src/components/Select/index.ts
@@ -1,0 +1,2 @@
+export * from './SelectSimple';
+export * from './CheckboxSelect';